# Speed up function template browsers in indirect analysis GUI