This PR speeds up/ improves the responsiveness of the function browser interaction in the indirect analysis GUI. The unresponsiveness was noted while trying to implement default parameter estimates within the MSD tab and therefore warranted an investigation. The primary issues found were:

-	When a function was changed in the browser the parameters were updated, with each parameter update sending a Qt signal.
-	There are a lot of replot/redraw calls. For instance, changing the function in the browser requested 5 replots/redraws of the graph, despite the data not changing. Redrawing the graph in Matplotlib is a fairly costly procedure.


Some timing results of changing the function in the MSD tab (on my Debug build), after the first and second set of changes.


  | Without data loaded –time (ms) | With data loaded - time (ms)
-- | -- | --
Master | 587 | 1068
Sending Qt signals silently | 108 | 190
Removing excess redrawing | 50 | 88



**To test:**
- Data to test the GUI can be found in #26631
- Run through each tab, testing that the functionality still works.

*There is no associated issue.*

---

#### Reviewer ####

Please comment on the following ([full description](http://developer.mantidproject.org/ReviewingAPullRequest.html)):

##### Code Review #####

- Is the code of an acceptable quality?
- Does the code conform to the [coding standards](http://developer.mantidproject.org/Standards/)?
- Are the unit tests small and test the class in isolation?
- If there is GUI work does it follow the [GUI standards](http://developer.mantidproject.org/Standards/GUIStandards.html)?
- If there are changes in the release notes then do they describe the changes appropriately?

##### Functional Tests #####

- Do changes function as described? Add comments below that describe the tests performed?
- Do the changes handle unexpected situations, e.g. bad input?
- Has the relevant (user and developer) documentation been added/updated?

Does everything look good? Mark the review as **Approve**. A member of `@mantidproject/gatekeepers` will take care of it.